requirement.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.50.0
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
macholib==1.16.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.5.8
numpy==1.26.1
openpyxl==3.1.5
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.9
WooCommerce==3.0.0
yarl==1.9.4
zipp==3.17.0
